Moving e2e, spec files under test folder
