+++ src/components/app-profile/test/app-profile.spec.ts
@@ -1,4 +1,4 @@
-import { AppProfile } from './app-profile';
+import { AppProfile } from '../app-profile';
 import { newSpecPage } from '@stencil/core/testing';
 
 describe('app-profile', () => {
